chore: 1.Added coredata stack 2. Added Recent search model. 3. Added Movie search controller.

+++ MovieApp/MovieApp/Home/Model/MovieListResponse.swift
@@ -15,7 +15,7 @@ struct MovieListResponse: Codable {
 }
 
 struct Movie: Codable {
-    let id: Int?
+    let id: Int64?
     let title: String?
     let overview: String?
     let voteAverage: Float?
